backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
python-dotenv>=1.0
django-cors-headers>=4.4
